add colour to decrypt_encrypt.py

+++ data/decrypt_encrypt.py
@@ -4,6 +4,7 @@ from cryptography.fernet import Fernet
 import os
 from __init__ import Config
 import json
+import colorama
 
 parser = argparse.ArgumentParser(description='Decrypt and encrypt data files for backend')
 parser.add_argument('-d', '--decrypt', help='Decrypt data files', action='store_true')
@@ -12,34 +13,46 @@ args = parser.parse_args()
 
 key = base64.urlsafe_b64encode(Config.JSON_SECRET_KEY.encode())
 fernet = Fernet(key)
-
-if args.decrypt and args.encrypt:
-    print("You can only decrypt or encrypt, not both")
-    exit(1)
-elif args.decrypt:
-    print("Decrypting data files")
-    for filename in os.listdir():
-        if filename.endswith(".json"):
-            with open(filename, "rb+") as f:
-                data = f.read().decode()
-                data = json.loads(data)["encrypted"]
-                decrypted = fernet.decrypt(data.encode()).decode()
-                f.seek(0)
-                f.write(json.dumps(json.loads(decrypted), indent=4).encode())
-                f.truncate()
-                f.close()
-    exit(0)
-elif args.encrypt:
-    print("Encrypting data files")
-    for filename in os.listdir():
-        if filename.endswith(".json"):
-            with open(filename, "rb+") as f:
-                data = json.load(f)
-                encrypted = json.dumps(data).encode()
-                encrypted = fernet.encrypt(encrypted).decode()
-                write = {"encrypted": encrypted}
-                f.seek(0)
-                f.write(json.dumps(write, indent=4).encode())
-                f.truncate()
-                f.close()
-    exit(0)
+colorama.init()
+try:
+    if not args.decrypt and not args.encrypt:
+        print(colorama.Fore.RED + " :: Please specify either --decrypt or --encrypt")
+        exit(1)
+    if args.decrypt and args.encrypt:
+        print(colorama.Fore.RED + " :: You can only decrypt or encrypt, not both")
+        exit(1)
+    elif args.decrypt:
+        try:
+            print(colorama.Fore.GREEN + " :: Decrypting data files")
+            for filename in os.listdir():
+                if filename.endswith(".json"):
+                    with open(filename, "rb+") as f:
+                        data = f.read().decode()
+                        data = json.loads(data)["encrypted"]
+                        decrypted = fernet.decrypt(data.encode()).decode()
+                        f.seek(0)
+                        f.write(json.dumps(json.loads(decrypted), indent=4).encode())
+                        f.truncate()
+                        f.close()
+            exit(0)
+        except KeyError:
+            print(colorama.Fore.RED + " :: Data is already decrypted")
+    elif args.encrypt:
+        print(colorama.Fore.GREEN + " :: Encrypting data files")
+        for filename in os.listdir():
+            if filename.endswith(".json"):
+                with open(filename, "rb+") as f:
+                    data = json.load(f)
+                    if data.get("encrypted"):
+                        print(colorama.Fore.RED + " :: Data is already encrypted")
+                        exit(0)
+                    encrypted = json.dumps(data).encode()
+                    encrypted = fernet.encrypt(encrypted).decode()
+                    write = {"encrypted": encrypted}
+                    f.seek(0)
+                    f.write(json.dumps(write, indent=4).encode())
+                    f.truncate()
+                    f.close()
+        exit(0)
+finally:
+    colorama.Style.RESET_ALL
